Fix rc.local for boostrap based on CentOS 7
Systemd reports failed to start /etc/rc.d/rc.local compatibility, this
commit fixes the problem.

Devpts is already mounted on /dev/pts on CentOS 7, so simply don't do
that again in the rc.local script.

For flock, "-w 0" is no longer a synonym for "-n" in the version we now
use on CentOS 7.

Fixes: redmine #3202

Signed-off-by: huntxu <mhuntxu@gmail.com>